apisComplete

+++ api/views.py
@@ -5,9 +5,22 @@ from django.http import HttpResponse, JsonResponse
 from rest_framework.response import Response
 import jwt
 from .models import *
+from django.contrib.auth import authenticate, login, logout, update_session_auth_hash
+from datetime import datetime, timedelta
+
+def validateJWT(request):
+    jwtToken = request.META['HTTP_AUTHORIZATION']
+    try:
+        validation = jwt.decode(jwtToken, 'PaulPogba', algorithms="HS256")
+        return True
+    except:
+        return False
+
 
 class Shows(APIView):
     def get(self, request):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         allShows = Show.objects.all()
         allTheShows = []
         for theShow in allShows:
@@ -23,6 +36,8 @@ class Shows(APIView):
 
 class TheShow(APIView):
     def get(self, request, showID):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             theShow = Show.objects.filter(id=showID)[0]
             thatShow = {
@@ -46,7 +61,10 @@ class TheShow(APIView):
 
 
 class TheEpisodeData(APIView):
+
     def get(self, request, episodeID):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             theEpisode = Episode.objects.filter(id=episodeID)[0]
             theComments = Comment.objects.filter(episode=theEpisode)
@@ -95,6 +113,8 @@ class TheEpisodeData(APIView):
 class CommentAPI(APIView):
     def post(self, request, episodeID, username):
         
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             thatEpisode = Episode.objects.filter(id=episodeID)[0]
         except:
@@ -124,7 +144,11 @@ class CommentAPI(APIView):
 
 
 class ReplyAPI(APIView):
+    
     def post(self,request,commentID,username):
+        
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             thatComment = Comment.objects.filter(id=commentID)[0]
         except:
@@ -151,6 +175,8 @@ class ReplyAPI(APIView):
 
 class theComment(APIView):
     def delete(self, request, id):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         comment = Comment.objects.filter(id=id)
         if comment:
             comment.delete()
@@ -159,6 +185,8 @@ class theComment(APIView):
        
     
     def put(self, request, id):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             text = request.data["text"]
             comment = Comment.objects.filter(id=id).update(text=text)
@@ -169,7 +197,10 @@ class theComment(APIView):
 
 class theReply(APIView):
     def delete(self, request, id):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         #Boooooooooo
+
         reply = Reply.objects.filter(id=id)
         if reply:
             reply.delete()
@@ -177,9 +208,101 @@ class theReply(APIView):
         return Response({"status": "404 Not Found", "message": "Reply does not exist."})
         
     def put(self, request, id):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
         try:
             text = request.data["text"]
             reply = Reply.objects.filter(id=id).update(text=text)
             return Response({"status":"Reply Updated successfully"})
         except: 
             return Response({"status": "404 Not Found", "message": "Reply does not exist."})
+
+class LoginAPI(APIView):
+
+    def post(self, request):
+        JWT_SECRET = 'PaulPogba'
+        JWT_ALGORITHM = 'HS256'
+        # JWT_EXP_DELTA_SECONDS = 2628000
+        JWT_EXP_DELTA_SECONDS = 2628000
+        username = request.data['username']
+        password = request.data['password']
+        user = authenticate(username=username, password=password)
+        if user is not None:
+            payload = {
+                'exp': datetime.utcnow() + timedelta(seconds=JWT_EXP_DELTA_SECONDS)
+                }
+            
+            jwt_token = jwt.encode(payload, JWT_SECRET, JWT_ALGORITHM)
+            
+            return Response({"status": "200 OK", "username": username, "token": jwt_token})
+        else:
+            return Response({"status": "400 Bad Request", "message": "Invalid Password/Username"})
+
+
+class RegisterAPI(APIView):
+    def post(self, request):
+        JWT_SECRET = 'PaulPogba'
+        JWT_ALGORITHM = 'HS256'
+        JWT_EXP_DELTA_SECONDS = 2628000
+        username = request.data['username']
+        password = request.data['password']
+        email = request.data['email']
+        fname = request.data['fname']
+        lname = request.data['lname']
+        
+        try:
+            user = User.objects.create_user(username, email, password, first_name=fname,last_name=lname)
+            user.save()
+            try:
+                payload = {
+                'user_id': user.id,
+                'exp': datetime.utcnow() + timedelta(seconds=JWT_EXP_DELTA_SECONDS)
+                }
+            
+                jwt_token = jwt.encode(payload, JWT_SECRET, JWT_ALGORITHM)
+                
+                return Response({"status": "200 OK", "username": username, "fname": fname, "lname": lname, "email": email, "token": jwt_token})
+            except:
+                return Response({"status": "400 Bad Request", "message": "Invalid Password/Username"})
+        except:
+            return Response({"status": "403 User already exists", "message": "User already exists."})
+
+class ProfileAPI(APIView):
+    def get(self, request, username):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
+        try:
+            thatUser = User.objects.filter(username=username)[0]
+            resp = {
+                "status": "200 OK",
+                "username": username,
+                "first_name": thatUser.first_name,
+                "last_name": thatUser.last_name,
+                "email": thatUser.email
+            }
+            return Response(resp)
+        except:
+            return Response({"status": "404 Not Found", "message": "username does not exist."})
+        
+    def put(self, request, username):
+        if validateJWT(request) is False:
+            return Response({"status": "401 Unauthorized", "message": "authentication token invalid."})
+        try:
+            email = request.data['email']
+            first_name = request.data['first_name']
+            last_name = request.data['last_name']
+            if len(User.objects.filter(username=username)) == 0:
+                return Response({"status": "404 Not Found", "message": "username does not exist."})
+            User.objects.filter(username=username).update(email=email, first_name=first_name, last_name=last_name)
+            thatUser = User.objects.filter(username=username)[0]
+            resp = {
+                "status": "200 OK",
+                "message": "successfully updated.",
+                "username": username,
+                "first_name": thatUser.first_name,
+                "last_name": thatUser.last_name,
+                "email": thatUser.email
+            }
+            return Response(resp)
+        except:
+            return Response({"status": "404 Not Found", "message": "username does not exist."})
